requirements-dev.txt
pytest
pytest-django
pytest-cov
coverage
